HPAdaptEditPage: add button to set defaults
This adds the possibility to add "Set Defaults" or "Reset" buttons to
any EditPageGeneric.  For now, a "Fill empty Resources" button that sets
all data resources has been added to HPAdaptEdit page.  This is handy
for testing and development, but in the future, it may be desirable to
only set defaults for some values.