EditPageGeneric: Fix limit check for temperatures
Instead of comparing raw float values, compare Celsius and
also assuming that limits are given in Celsius.

This is in accordance with the behavior of float-edit text fields
defined in `ObjectResourceGUIHelper.floatEdit(...)`, where Celsius is
always assumed as the unit of user-entered temperature values.

This fixes data containing negative Celsius temperature values
erroneously failing the resource check, prohibiting its activation.